// frontend/src/api/model/githubCom4H1RZooraInternalDomainAIGradingJob.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainAIGradingMode } from './githubCom4H1RZooraInternalDomainAIGradingMode';
import type { GithubCom4H1RZooraInternalDomainAIGradingStatus } from './githubCom4H1RZooraInternalDomainAIGradingStatus';

export interface GithubCom4H1RZooraInternalDomainAIGradingJob {
  created_at?: string;
  created_by?: string;
  done?: number;
  error?: string;
  failed?: number;
  id?: string;
  mode?: GithubCom4H1RZooraInternalDomainAIGradingMode;
  organization_id?: string;
  quiz_id?: string;
  status?: GithubCom4H1RZooraInternalDomainAIGradingStatus;
  total?: number;
  updated_at?: string;
}